requirements.txt
flask==2.3.2
flask-cors==3.0.10
gunicorn==21.2.0
transformers==4.40.1
reportlab==4.0.6
google-cloud-firestore==2.11.0
requests==2.31.0
python-dotenv==1.0.0
